refactor: LottoNumbers 클래스 Lotto 클래스로 통합
